# onewire.py: Optimize timing, enable CRC check and slim the code