@amotl 2019-11-17
For everyone reading this, @robert-hh just added the improvements to the pure-Python variant to a separate repository [1].

Apart from that, we have been able to unlock the native 1-Wire/DS18X20 variant coming from Genuine MicroPython through [2].

[1] https://github.com/robert-hh/Onewire_DS18X20
[2] https://github.com/pycom/pycom-micropython-sigfox/pull/356